requirements.txt
accelerate==0.34.2
albucore==0.0.15
albumentations==1.4.15
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
easydict==1.13
ecdsa==0.19.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
fastapi==0.114.2
filelock==3.16.1
flake8==7.1.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
iniconfig==2.0.0
insightface==0.7.3
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mpmath==1.3.0
mypy==1.11.2
mypy-extensions==1.0.0
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
onnx==1.16.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
pluggy==1.5.0
prettytable==3.11.0
protobuf==5.28.1
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pyflakes==3.2.0
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
regex==2024.5.15
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
soundfile==0.12.1
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.5
sympy==1.13.2
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
httpx